+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @7ito/sketcherson-common
 
+## 0.3.19
+
+### Patch Changes
+
+- Emotes fix
+
 ## 0.3.18
 
 ### Patch Changes