+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @7ito/sketcherson-server
 
+## 0.3.19
+
+### Patch Changes
+
+- Emotes fix
+- Updated dependencies
+  - @7ito/sketcherson-common@0.3.19
+
 ## 0.3.18
 
 ### Patch Changes